Добавлены заголовочные файлы для графики и окна, реализована основная логика приложения Wi-Fi Radar, удален устаревший скрипт checkadapter.py, Разделены модули, настроены параметры триангуляции
