+++ .vscode/settings.json
@@ -54,5 +54,44 @@
         "xmemory": "cpp",
         "xtr1common": "cpp",
         "xutility": "cpp"
-    }
+    },
+    "cSpell.words": [
+        "Идентификаторы",
+        "элементов",
+        "управления",
+        "Таймер",
+        "сонара",
+        "Создание",
+        "кисти",
+        "фона",
+        "списка",
+        "точек",
+        "доступа",
+        "координаты",
+        "кнопки",
+        "сохранения",
+        "расчета",
+        "азимута",
+        "мерцания",
+        "триангуляции",
+        "радианах",
+        "PAINTSTRUCT",
+        "HBITMAP",
+        "HGDIOBJ",
+        "HINSTANCE",
+        "WNDCLASSW",
+        "lpfn",
+        "lpsz",
+        "HREDRAW",
+        "OWNDC",
+        "OVERLAPPEDWINDOW",
+        "CLIPCHILDREN",
+        "wlanapi",
+        "LPCCH",
+        "rssi",
+        "RSSI",
+        "BSSID",
+        "Bssid",
+        "wtof"
+    ]
 }
