+++ window.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <windows.h>
+
+LRESULT CALLBACK WindowProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
